codecoverage/backend: Move gcp to main module
