+++ src/codecoverage/backend/codecoverage_backend/api.py
@@ -7,7 +7,7 @@ from flask import abort
 
 from cli_common import log
 from codecoverage_backend.config import COVERAGE_EXTENSIONS
-from codecoverage_backend.services.gcp import load_cache
+from codecoverage_backend.gcp import load_cache
 
 DEFAULT_REPOSITORY = 'mozilla-central'
 logger = log.get_logger(__name__)
